requirements.txt
numpy
pygame
Pillow
pymunk
inputs
